Fix tests and move ConfigPage component
