+++ apps/ai-content-generator/src/components/config/config-page/ConfigPage.styles.ts
