+++ apps/ai-content-generator/src/locations/ConfigScreen.tsx
@@ -1,4 +1,4 @@
-import ConfigPage from '@components/config/ConfigPage';
+import ConfigPage from '@components/config/config-page/ConfigPage';
 
 interface AppInstallationParameters {
   model: string;
